pyproject.toml
[project]
name = "recommend-site"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.11"
dependencies = [
    "bitsandbytes>=0.46.0",
    "bs4>=0.0.2",
    "cryptography>=45.0.4",
    "datasets>=3.6.0",
    "fake-agent>=0.1.4",
    "fake-useragent>=2.2.0",
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "lxml>=5.4.0",
    "lzstring>=1.0.4",
    "mysql-connector-python>=9.3.0",
    "notebook>=7.4.3",
    "pandas>=2.3.0",
    "parmap>=1.7.0",
    "peft>=0.15.2",
    "playwright>=1.52.0",
    "pymysql>=1.1.1",
    "requests>=2.32.3",
    "rsa>=4.9.1",
    "selenium>=4.33.0",
    "tqdm>=4.67.1",
    "trl>=0.19.0",
]
